requirements.txt
ffmpeg-python==0.2.0
pyinstaller==4.10
PySimpleGUI==4.57.0
